Spam filter added
If the user moves a mail into the spam folder, the sender's address is added to a blacklist; if the user moves a mail off the spam folder, the address is removed from the blacklist.

Everytime a new mail is detected in the INBOX folder, the sender's address is compared to the blacklist and as spam identified mails are moved to the spam folder, without showing a notification for new emails.